Extract pflag-shim helpers to builtins/internal/flagparser/
Pure refactor — no behaviour change. Moves the error-rewriting and
trial-parse helpers out of builtins.go into a dedicated package so
the file is back to focusing on builtin registration.

What moved:
  - rewritePflagError  → flagparser.RewriteError
  - trialHelpTrimIndex → flagparser.TrialHelpTrimIndex
  - extractFlagDescriptor, longFlagName, recoverLongFlagToken,
    shortFlagEqualsValueIn, shortFlagFromDescriptor, shortMissingArg
    (all unexported helpers)

builtins.go now imports flagparser and delegates. The trial-parse
helper's signature changed from `func(*FlagSet) HandlerFunc` to
`func(*FlagSet)` so the package doesn't need to know about
builtins.HandlerFunc — the caller wraps factory in a one-line
closure that discards the returned handler.

Tests moved with the code (builtins/internal/flagparser/flagparser_test.go).
analysis/symbols_internal.go gained a flagparser entry plus
io.Discard, strings.CutPrefix, strings.HasSuffix, and the three
pflag symbols used by the trial parser.

Addresses [3204000991](https://github.com/DataDog/rshell/pull/238#discussion_r3204000991)
option #2. Keeps option #1 (own parser) as a future possibility:
if we ever swap pflag for a hand-written GNU-getopt implementation,
the swap is local to this package.
